Add asteroid torque
Replace files with new tested versions
