+++ Assets/Materials.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 0f897778ba2cf01488bfbf017ca02f1b
-PrefabImporter:
+guid: 01fb011cde7b7314c80fb90c6b867d72
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
