+++ Assets/Materials/GalaxyFront.mat.meta
