+++ Assets/Prefabs/Asteroid.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &7410811981543854193
+--- !u!1 &5136191482376294984
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,11 +8,11 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 6698849281816595175}
-  - component: {fileID: 2559035984918543218}
-  - component: {fileID: 311774397064228452}
-  - component: {fileID: 6040064341035749879}
-  - component: {fileID: 3250687757763649870}
+  - component: {fileID: 5829112507523505775}
+  - component: {fileID: 8274631640572479202}
+  - component: {fileID: 2795835151576491281}
+  - component: {fileID: 862605777385663934}
+  - component: {fileID: 883024211555833721}
   m_Layer: 0
   m_Name: Asteroid
   m_TagString: Asteroid
@@ -20,13 +20,13 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &6698849281816595175
+--- !u!4 &5829112507523505775
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7410811981543854193}
+  m_GameObject: {fileID: 5136191482376294984}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 0.5, y: 0.5, z: 1}
@@ -34,13 +34,13 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!212 &2559035984918543218
+--- !u!212 &8274631640572479202
 SpriteRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7410811981543854193}
+  m_GameObject: {fileID: 5136191482376294984}
   m_Enabled: 1
   m_CastShadows: 0
   m_ReceiveShadows: 0
@@ -84,14 +84,14 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!50 &311774397064228452
+--- !u!50 &2795835151576491281
 Rigidbody2D:
   serializedVersion: 4
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7410811981543854193}
+  m_GameObject: {fileID: 5136191482376294984}
   m_BodyType: 0
   m_Simulated: 1
   m_UseFullKinematicContacts: 0
@@ -105,110 +105,31 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 1
   m_Constraints: 0
---- !u!60 &6040064341035749879
-PolygonCollider2D:
+--- !u!58 &862605777385663934
+CircleCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7410811981543854193}
+  m_GameObject: {fileID: 5136191482376294984}
   m_Enabled: 1
   m_Density: 1
   m_Material: {fileID: 0}
   m_IsTrigger: 1
   m_UsedByEffector: 0
   m_UsedByComposite: 0
-  m_Offset: {x: 0, y: 0}
-  m_SpriteTilingProperty:
-    border: {x: 0, y: 0, z: 0, w: 0}
-    pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 2.56, y: 2.56}
-    newSize: {x: 2.56, y: 2.56}
-    adaptiveTilingThreshold: 0.5
-    drawMode: 0
-    adaptiveTiling: 0
-  m_AutoTiling: 0
-  m_Points:
-    m_Paths:
-    - - {x: 0.5876005, y: -0.4556779}
-      - {x: 0.6054718, y: -0.4272677}
-      - {x: 0.65436506, y: -0.37701517}
-      - {x: 0.7118253, y: -0.30635923}
-      - {x: 0.74273396, y: -0.23454921}
-      - {x: 0.7754563, y: -0.20091534}
-      - {x: 0.78909504, y: -0.16181691}
-      - {x: 0.78621364, y: -0.05354748}
-      - {x: 0.81501, y: 0.017362222}
-      - {x: 0.8318213, y: 0.0881279}
-      - {x: 0.8136466, y: 0.1518169}
-      - {x: 0.78182924, y: 0.20908463}
-      - {x: 0.71091646, y: 0.26275986}
-      - {x: 0.70514727, y: 0.31825343}
-      - {x: 0.68496084, y: 0.3367406}
-      - {x: 0.6962784, y: 0.36294523}
-      - {x: 0.6939443, y: 0.4009065}
-      - {x: 0.6881903, y: 0.43545768}
-      - {x: 0.6409699, y: 0.4768361}
-      - {x: 0.63372016, y: 0.51241726}
-      - {x: 0.5890951, y: 0.5654645}
-      - {x: 0.5682603, y: 0.6087512}
-      - {x: 0.55455124, y: 0.67895854}
-      - {x: 0.5454641, y: 0.72454226}
-      - {x: 0.5272659, y: 0.73183763}
-      - {x: 0.4811757, y: 0.73670805}
-      - {x: 0.4453214, y: 0.73912567}
-      - {x: 0.4172738, y: 0.7409291}
-      - {x: 0.3505564, y: 0.75465065}
-      - {x: 0.22816694, y: 0.7536476}
-      - {x: 0.19392894, y: 0.74257773}
-      - {x: 0.16061991, y: 0.7317154}
-      - {x: 0.10972722, y: 0.7160591}
-      - {x: 0.009616286, y: 0.7196994}
-      - {x: -0.0022887886, y: 0.7057144}
-      - {x: -0.05406031, y: 0.69542027}
-      - {x: -0.11027469, y: 0.69443756}
-      - {x: -0.1672699, y: 0.70637304}
-      - {x: -0.2818097, y: 0.7536407}
-      - {x: -0.31727377, y: 0.7481968}
-      - {x: -0.43577832, y: 0.69933033}
-      - {x: -0.48980144, y: 0.61053}
-      - {x: -0.659186, y: 0.38598222}
-      - {x: -0.7065471, y: 0.3048905}
-      - {x: -0.7199954, y: 0.2468557}
-      - {x: -0.7498571, y: 0.20088032}
-      - {x: -0.807926, y: 0.09127468}
-      - {x: -0.8334076, y: 0.012440506}
-      - {x: -0.8458519, y: -0.047813118}
-      - {x: -0.8362446, y: -0.09449676}
-      - {x: -0.8199263, y: -0.12933038}
-      - {x: -0.77792585, y: -0.17224687}
-      - {x: -0.73738533, y: -0.26248667}
-      - {x: -0.703901, y: -0.3944641}
-      - {x: -0.6312601, y: -0.524881}
-      - {x: -0.60602766, y: -0.5478166}
-      - {x: -0.56133366, y: -0.5848164}
-      - {x: -0.40051964, y: -0.64585316}
-      - {x: -0.32725906, y: -0.6526785}
-      - {x: -0.2479259, y: -0.7103025}
-      - {x: -0.12881324, y: -0.7268476}
-      - {x: -0.002370689, y: -0.77377975}
-      - {x: 0.02711197, y: -0.75097215}
-      - {x: 0.1853086, y: -0.74410677}
-      - {x: 0.3223204, y: -0.72334117}
-      - {x: 0.3736233, y: -0.7063937}
-      - {x: 0.42726994, y: -0.6327667}
-      - {x: 0.47341868, y: -0.6043567}
-      - {x: 0.56006247, y: -0.510452}
-      - {x: 0.55909514, y: -0.48093602}
---- !u!114 &3250687757763649870
+  m_Offset: {x: -0.0047332644, y: 0.01786542}
+  serializedVersion: 2
+  m_Radius: 0.78575295
+--- !u!114 &883024211555833721
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7410811981543854193}
+  m_GameObject: {fileID: 5136191482376294984}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 069c7ba5ce0243941969fd0e1cbf748a, type: 3}
+  m_Script: {fileID: 11500000, guid: 89643dab3ed0a5e42adc59ccb5c9b2b1, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
