+++ Assets/Prefabs/Asteroid.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3310cf874b2f5f240a0df8f574884367
+guid: cca2f8a06665f7d4883818e46e57df62
 PrefabImporter:
   externalObjects: {}
   userData:
