+++ Assets/Prefabs/SpaceShip.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d25aed08691918f488b423405548b5a0
+guid: 796c71424336c31459b03dc201fb1996
 PrefabImporter:
   externalObjects: {}
   userData:
