+++ Assets/Scripts/Asteroid/Asteroid.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 069c7ba5ce0243941969fd0e1cbf748a
+guid: 89643dab3ed0a5e42adc59ccb5c9b2b1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
