+++ Assets/Scripts/Background.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: d98823ef427c56f48b9a3d1aee622b1f
-PrefabImporter:
+guid: 98ca5073a8a52604f863c9cb482b56f4
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
