--- Assets/Testing/AsteroidTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 89643dab3ed0a5e42adc59ccb5c9b2b1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
