--- Assets/Testing/Background.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fc9f8e81101f0774286d2aa858061e1e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
