--- Assets/Testing/Background/GalaxyDiffuseBack.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: babc6322c4ae9dd468cd7af415ec45b5
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
