--- Assets/Testing/Background/GalaxyDiffuseFront.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9c766ec32a3e5c8418e1ba782e1f529a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
