--- Assets/Testing/TestSpawn.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ed19be08bf8ef9a45ad38848fa5af563
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
