+++ Assets/Textures.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d959857b56409bf47b9d3c742b20964f
+guid: 8ae9cb7ed8f27b74e808c96ca0f8a325
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
