+++ Assets/Textures/GalaxySpace.PNG.meta
